+++ prometheus/histogram.go
@@ -430,7 +430,7 @@ func NewConstHistogram(
 	buckets map[float64]uint64,
 	labelValues ...string,
 ) (Metric, error) {
-	if err := validateLabelValues(labelValues, len(desc.variableLabels)); err != nil {
+	if err := validateValuesInLabels(labelValues, len(desc.variableLabels)); err != nil {
 		return nil, err
 	}
 	return &constHistogram{
