+++ prometheus/value.go
@@ -14,7 +14,6 @@
 package prometheus
 
 import (
-	"errors"
 	"fmt"
 	"math"
 	"sort"
@@ -36,8 +35,6 @@ const (
 	GaugeValue
 	UntypedValue
 )
-
-var errInconsistentCardinality = errors.New("inconsistent label cardinality")
 
 // value is a generic metric for simple values. It implements Metric, Collector,
 // Counter, Gauge, and Untyped. Its effective type is determined by
@@ -158,7 +155,7 @@ func (v *valueFunc) Write(out *dto.Metric) error {
 // the Collect method. NewConstMetric returns an error if the length of
 // labelValues is not consistent with the variable labels in Desc.
 func NewConstMetric(desc *Desc, valueType ValueType, value float64, labelValues ...string) (Metric, error) {
-	if err := validateLabelValues(labelValues, len(desc.variableLabels)); err != nil {
+	if err := validateValuesInLabels(labelValues, len(desc.variableLabels)); err != nil {
 		return nil, err
 	}
 	return &constMetric{
